Bubblesort using a singly linked list.
This implementation of bubblesort uses a singly linked list in order to
arrange the given elements (only integers placed in a singly linked
list) in increasing order. Note, however that this method uses a
timecomplexity that have been tuned to match the regular bubblesort
implementation:s timecomplexity. This version has many limitations of
usage, use on your own risk.